requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiomysql==0.2.0
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.10.0
attrs==25.1.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.2.1
discord==2.3.2
discord.py==2.4.0
fastapi==0.116.1
frozenlist==1.5.0
google-api-core==2.24.1
google-api-python-client==2.163.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.69.1
gspread==6.2.0
h11==0.16.0
httplib2==0.22.0
httptools==0.6.4
idna==3.10
multidict==6.1.0
mysql-connector-python==9.2.0
oauth2client==4.1.3
oauthlib==3.2.2
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.11.7
pydantic_core==2.33.2
PyMySQL==1.1.1
pyparsing==3.2.1
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
yarl==1.18.3
